Move Case Management integration test to showcase module
 Reuse workbench deployment to test case management provisioning API
