--- jbpm-wb-integration/jbpm-wb-integration-backend/src/test/resources/arquillian.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-            xmlns="http://jboss.org/schema/arquillian"
-            xsi:schemaLocation="http://jboss.org/schema/arquillian http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
-
-  <!-- Use Servlet 3.0 -->
-  <defaultProtocol type="Servlet 3.0"/>
-
-  <container qualifier="wildfly-managed" default="true">
-    <configuration>
-      <property name="jbossHome">${jboss.home}</property>
-      <property name="javaVmArguments">-Xmx512m -XX:MaxPermSize=128m -Dorg.uberfire.async.executor.safemode=true -Dorg.jbpm.casemgmt.showcase.deploy=true</property>
-    </configuration>
-  </container>
-</arquillian>
